[refactor] convertForecastViewModel 로직 분리
